+++ integration/auth/user_mgt_test.go
@@ -16,6 +16,7 @@
 package auth
 
 import (
+	"context"
 	"fmt"
 	"reflect"
 	"testing"
@@ -24,8 +25,6 @@ import (
 	"google.golang.org/api/iterator"
 
 	"firebase.google.com/go/auth"
-
-	"golang.org/x/net/context"
 )
 
 var testFixtures = struct {
@@ -371,6 +370,9 @@ func testRemoveCustomClaims(t *testing.T) {
 		t.Fatal(err)
 	}
 	u, err = client.GetUser(context.Background(), testFixtures.sampleUserBlank.UID)
+	if err != nil {
+		t.Fatal(err)
+	}
 	if u.CustomClaims != nil {
 		t.Errorf("CustomClaims() = %#v; want = nil", u.CustomClaims)
 	}