+++ internal/internal.go
@@ -21,6 +21,16 @@ import (
 	"google.golang.org/api/option"
 )
 
+// FirebaseScopes is the set of OAuth2 scopes used by the Admin SDK.
+var FirebaseScopes = []string{
+	"https://www.googleapis.com/auth/cloud-platform",
+	"https://www.googleapis.com/auth/datastore",
+	"https://www.googleapis.com/auth/devstorage.full_control",
+	"https://www.googleapis.com/auth/firebase",
+	"https://www.googleapis.com/auth/identitytoolkit",
+	"https://www.googleapis.com/auth/userinfo.email",
+}
+
 // AuthConfig represents the configuration of Firebase Auth service.
 type AuthConfig struct {
 	Opts      []option.ClientOption
@@ -33,6 +43,14 @@ type AuthConfig struct {
 type InstanceIDConfig struct {
 	Opts      []option.ClientOption
 	ProjectID string
+}
+
+// DatabaseConfig represents the configuration of Firebase Database service.
+type DatabaseConfig struct {
+	Opts         []option.ClientOption
+	URL          string
+	Version      string
+	AuthOverride map[string]interface{}
 }
 
 // StorageConfig represents the configuration of Google Cloud Storage service.