+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 -
 
+# v2.6.0
+
+- [changed] Improved error handling in FCM by mapping more server-side
+  errors to client-side error codes.
+- [added] Added the `db` package for interacting with the Firebase database.
+
 # v2.5.0
 
 - [changed] Import context from `golang.org/x/net` for 1.6 compatibility